requirements.txt
ollama
openai
python-dotenv
requests
beautifulsoup4